+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## 1.0.0 (2023-12-12)
 
-
 ### Features
 
-* added initdjango command to app ([649e249](https://github.com/djangoheads/djangoheads/commit/649e249230d84aea1f80dfc43b398ee1e369862c))
+* added initdjango command to
+  app ([649e249](https://github.com/djangoheads/djangoheads/commit/649e249230d84aea1f80dfc43b398ee1e369862c))